Move dataset_ops.py up a level (it's by itself in a directory) and remove unused imports
