+++ packages/grid/apps/domain/src/main/routes/data_centric/datasets/routes.py
@@ -6,13 +6,8 @@ from json import loads
 # third party
 from flask import Response
 from flask import request
-from main.core.datasets.dataset_ops import create_df_dataset
-from main.core.exceptions import AuthorizationError
+from main.core.dataset_ops import create_df_dataset
 from main.core.task_handler import route_logic
-from main.core.task_handler import task_handler
-from main.utils.executor import executor
-from syft.core.node.common.service.repr_service import ReprMessage
-from syft.grid.messages.dataset_messages import CreateDatasetMessage
 from syft.grid.messages.dataset_messages import DeleteDatasetMessage
 from syft.grid.messages.dataset_messages import GetDatasetMessage
 from syft.grid.messages.dataset_messages import GetDatasetsMessage
